requirements.txt
certifi==2021.5.30
chardet==4.0.0
cheroot==8.5.2
CherryPy==18.6.0
idna==2.10
jaraco.classes==3.2.1
jaraco.collections==3.3.0
jaraco.functools==3.3.0
jaraco.text==3.5.0
more-itertools==8.8.0
paho-mqtt==1.5.1
portend==2.7.1
pytz==2021.1
requests==2.25.1
six==1.16.0
tempora==4.1.1
urllib3==1.26.6
zc.lockfile==2.0
